+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## Breaking changes
 * Deprecated `Space` in favor of the individual spaces: `Nothing, GridSpace, GraphSpace, ContinuousSpace`.
-* The `n::Function` argument of `step!` now takes two arguments.
 * Reworked the public API of `GridSpace` to be simpler: position must be `NTuple{Int}`. As a result `vertex2coord` and stuff no longer exported, since they are obsolete.
 * It is now mandatory to provide a concrete type when initializing an `ABM` (for type stability and performance)
 
@@ -11,6 +10,7 @@
 * new function `space_neighbors`, which works for any space. It always and consistently returns the **IDs** of neighbors irrespectively
   of the spatial structure.
 * `when` of `step!` can now be `true`.
+* `step!` has a keyword `Nmax`, used only when `n::Function`, that terminates the evolution if the model steps more than `Nmax`, regardless of what `n(model)` outputs.
 
 # v2.1
 * Renamed the old scheduler `as_added` to `by_id`, to reflect reality.